move get_scene to utils
