+++ src/ansys/mechanical/core/embedding/viz/usd_converter.py
@@ -26,7 +26,7 @@ import typing
 
 from pxr import Gf, Usd, UsdGeom
 
-from .utils import bgr_to_rgb_tuple, get_nodes_and_coords
+from .utils import bgr_to_rgb_tuple, get_nodes_and_coords, get_scene
 
 
 def _transform_to_rotation_translation(
@@ -119,35 +119,6 @@ def _convert_attribute_node(
     child_node = node.Child
     color = node.Property(Ansys.Mechanical.Scenegraph.ScenegraphIntAttributes.Color)
     _convert_transform_node(child_node, stage, path, bgr_to_rgb_tuple(color))
-
-
-def get_scene(
-    app: "ansys.mechanical.core.embedding.App",
-) -> "Ansys.Mechanical.Scenegraph.GroupNode":
-    """Get the scene of the model"""
-    import clr
-
-    clr.AddReference("Ansys.Mechanical.DataModel")
-    clr.AddReference("Ansys.Mechanical.Scenegraph")
-    clr.AddReference("Ansys.ACT.Interfaces")
-
-    import Ansys  # isort: skip
-
-    category = Ansys.Mechanical.DataModel.Enums.DataModelObjectCategory.Body
-    group_node = Ansys.Mechanical.Scenegraph.Builders.GroupNodeBuilder()
-    for body in app.DataModel.GetObjectsByType(category):
-        scenegraph_node = Ansys.ACT.Mechanical.Tools.ScenegraphHelpers.GetScenegraph(body)
-        # wrap the body node in an attribute node using the body color
-        attribute_node_builder = Ansys.Mechanical.Scenegraph.Builders.AttributeNodeBuilder()
-        attribute_node = (
-            attribute_node_builder.Tag(f"body{body.ObjectId}")
-            .Child(scenegraph_node)
-            # set the color, body.Color is a BGR uint bitfield
-            .Property(Ansys.Mechanical.Scenegraph.ScenegraphIntAttributes.Color, body.Color)
-            .Build()
-        )
-        group_node.AddChild(attribute_node)
-    return group_node.Build()
 
 
 def load_into_usd_stage(app: "ansys.mechanical.core.embedding.App", stage: Usd.Stage) -> None:
